Replace Player Cube with Ship Model
I've got it set up so that the ship model is now the player's character (its position is controlled by the arrow keys) and it can collide with the boundaries and fire bullets. However, when the enemies collide with the ship, they vanish but do not push the ship back. Can someone fix this?